slim down GrpcReflection module by separating its supervisor
