+++ lib/grpc_reflection.ex
@@ -34,7 +34,6 @@ defmodule GrpcReflection do
   end
   ```
   """
-  use Supervisor
 
   @type descriptor_t ::
           %Google.Protobuf.DescriptorProto{} | %Google.Protobuf.ServiceDescriptorProto{}
@@ -130,35 +129,13 @@ defmodule GrpcReflection do
     end
   end
 
-  def start_link(opts) do
-    Supervisor.start_link(__MODULE__, opts, name: __MODULE__)
-  end
-
-  @impl true
-  def init(opts) do
-    children = [
-      %{
-        id: OneOffTask,
-        start: {Task, :start_link, [__MODULE__, :one_off_task, []]},
-        type: :worker,
-        restart: :temporary
-      },
-      %{
-        id: GrpcReflection.DynamicSupervisor,
-        start: {GrpcReflection.DynamicSupervisor, :start_link, [opts]},
-        type: :supervisor,
-        restart: :permanent
-      }
-    ]
-
-    Supervisor.init(children, strategy: :one_for_one)
-  end
-
-  @doc """
-  Load the protobuf extensions
-  """
-  def one_off_task do
-    Protobuf.load_extensions()
-    :ignore
+  @doc false
+  def child_spec(opts) do
+    %{
+      id: GrpcReflection,
+      start: {GrpcReflection.Supervisor, :start_link, [opts]},
+      type: :supervisor,
+      restart: :permanent
+    }
   end
 end
